+++ libraries/transformer/src/androidTest/java/androidx/media3/transformer/CompositionPlayerGapsTest.java
@@ -19,20 +19,20 @@ import static androidx.media3.common.util.Util.isRunningOnEmulator;
 import static androidx.media3.test.utils.AssetInfo.MP3_ASSET;
 import static androidx.media3.test.utils.AssetInfo.MP4_ASSET;
 import static com.google.common.truth.Truth.assertThat;
-import static org.junit.Assert.assertThrows;
 import static org.junit.Assume.assumeFalse;
 
 import android.app.Instrumentation;
 import android.content.Context;
 import android.view.SurfaceView;
+import androidx.media3.common.C;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.MimeTypes;
-import androidx.media3.common.PlaybackException;
 import androidx.media3.exoplayer.video.VideoFrameMetadataListener;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
+import com.google.common.collect.ImmutableSet;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.junit.After;
@@ -92,37 +92,8 @@ public final class CompositionPlayerGapsTest {
   }
 
   @Test
-  public void playback_withTwoMediaItemsAndGapAtStartWithForceVideoTrackSetToFalse_throws() {
-    PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
-    Composition composition =
-        new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceAudioTrack(true)
-                    .experimentalSetForceVideoTrack(false)
-                    .addGap(GAP_DURATION_US)
-                    .addItem(AUDIO_VIDEO_MEDIA_ITEM)
-                    .addItem(AUDIO_VIDEO_MEDIA_ITEM)
-                    .build())
-            .build();
-
-    instrumentation.runOnMainSync(
-        () -> {
-          compositionPlayer = new CompositionPlayer.Builder(context).build();
-          // Set a surface on the player even though there is no UI on this test. We need a surface
-          // otherwise the player will skip/drop video frames.
-          compositionPlayer.setVideoSurfaceView(surfaceView);
-          compositionPlayer.addListener(listener);
-          compositionPlayer.setComposition(composition);
-          compositionPlayer.prepare();
-          compositionPlayer.play();
-        });
-
-    assertThrows(PlaybackException.class, listener::waitUntilPlayerEnded);
-  }
-
-  @Test
   public void
-      playback_withTwoMediaItemsAndGapAtStartWithForceVideoTrackSetToTrue_rendersOneByOneBlackFramesForGap()
+      playback_withTwoMediaItemsAndGapAtStart_inAudioVideoSequence_rendersOneByOneBlackFramesForGap()
           throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
@@ -130,8 +101,8 @@ public final class CompositionPlayerGapsTest {
     long gapEndCompositionUs = GAP_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addGap(GAP_DURATION_US)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
@@ -173,16 +144,17 @@ public final class CompositionPlayerGapsTest {
   }
 
   @Test
-  public void playback_withTwoMediaItemsAndGapInMiddle_rendersOneByOneBlackFramesForGap()
-      throws Exception {
+  public void
+      playback_withTwoMediaItemsAndGapInMiddle_inAudioVideoSequence_rendersOneByOneBlackFramesForGap()
+          throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
     long gapStartCompositionUs = AUDIO_VIDEO_MEDIA_ITEM_DURATION_US;
     long gapEndCompositionUs = gapStartCompositionUs + GAP_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addGap(GAP_DURATION_US)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
@@ -225,16 +197,17 @@ public final class CompositionPlayerGapsTest {
   }
 
   @Test
-  public void playback_withTwoMediaItemsAndGapAtTheEnd_rendersOneByOneBlackFramesForGap()
-      throws Exception {
+  public void
+      playback_withTwoMediaItemsAndGapAtTheEnd_inAudioVideoSequence_rendersOneByOneBlackFramesForGap()
+          throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
     long gapStartCompositionUs = 2 * AUDIO_VIDEO_MEDIA_ITEM_DURATION_US;
     long gapEndCompositionUs = gapStartCompositionUs + GAP_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addGap(GAP_DURATION_US)
@@ -278,36 +251,7 @@ public final class CompositionPlayerGapsTest {
 
   @Test
   public void
-      playback_withThreeMediaItemsAndFirstMediaItemHavingNoVideoWithForceVideoTrackSetToFalse_throws() {
-    PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
-    Composition composition =
-        new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(false)
-                    .addItem(AUDIO_ONLY_MEDIA_ITEM)
-                    .addItem(AUDIO_VIDEO_MEDIA_ITEM)
-                    .addItem(AUDIO_VIDEO_MEDIA_ITEM)
-                    .build())
-            .build();
-
-    instrumentation.runOnMainSync(
-        () -> {
-          compositionPlayer = new CompositionPlayer.Builder(context).build();
-          // Set a surface on the player even though there is no UI on this test. We need a surface
-          // otherwise the player will skip/drop video frames.
-          compositionPlayer.setVideoSurfaceView(surfaceView);
-          compositionPlayer.addListener(listener);
-          compositionPlayer.setComposition(composition);
-          compositionPlayer.prepare();
-          compositionPlayer.play();
-        });
-
-    assertThrows(PlaybackException.class, listener::waitUntilPlayerEnded);
-  }
-
-  @Test
-  public void
-      playback_withThreeMediaItemsAndFirstMediaItemHavingNoVideoWithForceVideoTrackSetToTrue_rendersOneByOneBlackFramesForFirstMediaItem()
+      playback_withThreeMediaItemsAndFirstMediaItemHavingNoVideo_inAudioVideoSequence_rendersOneByOneBlackFramesForFirstMediaItem()
           throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
@@ -315,8 +259,8 @@ public final class CompositionPlayerGapsTest {
     long audioOnlyItemEndCompositionUs = AUDIO_ONLY_MEDIA_ITEM_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addItem(AUDIO_ONLY_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
@@ -359,7 +303,7 @@ public final class CompositionPlayerGapsTest {
 
   @Test
   public void
-      playback_withThreeMediaItemsAndSecondMediaItemHavingNoVideo_rendersOneByOneBlackFramesForSecondMediaItem()
+      playback_withThreeMediaItemsAndSecondMediaItemHavingNoVideo_inAudioVideoSequence_rendersOneByOneBlackFramesForSecondMediaItem()
           throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
@@ -368,8 +312,8 @@ public final class CompositionPlayerGapsTest {
         audioOnlyItemStartCompositionUs + AUDIO_ONLY_MEDIA_ITEM_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_ONLY_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
@@ -413,7 +357,7 @@ public final class CompositionPlayerGapsTest {
 
   @Test
   public void
-      playback_withThreeMediaItemsAndLastMediaItemHavingNoVideo_rendersOneByOneBlackFramesForLastMediaItem()
+      playback_withThreeMediaItemsAndLastMediaItemHavingNoVideo_inAudioVideoSequence_rendersOneByOneBlackFramesForLastMediaItem()
           throws Exception {
     assumeFalse("Skipped on emulator due to surface dropping frames", isRunningOnEmulator());
     PlayerTestListener listener = new PlayerTestListener(TEST_TIMEOUT_MS);
@@ -422,8 +366,8 @@ public final class CompositionPlayerGapsTest {
         audioOnlyItemStartCompositionUs + AUDIO_ONLY_MEDIA_ITEM_DURATION_US;
     Composition composition =
         new Composition.Builder(
-                new EditedMediaItemSequence.Builder()
-                    .experimentalSetForceVideoTrack(true)
+                new EditedMediaItemSequence.Builder(
+                        ImmutableSet.of(C.TRACK_TYPE_AUDIO, C.TRACK_TYPE_VIDEO))
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_VIDEO_MEDIA_ITEM)
                     .addItem(AUDIO_ONLY_MEDIA_ITEM)